make-xges: add a flag to stretch webcam footage to 16:9
This is to work around a quirk with the HDMI capture card we've been
using at PLUG.  BigBlueButton seems to always ask for a 480 line frame
size, which ends up picking the 640x480 mode.  This results in the
capture card returning the widescreen HDMI stream squashed to 4:3.

This option stretches the video back out to the original aspect ratio.