make-xges: use _add_clip helper to add opening/closing credits
A duration can be set for each card.  If no duration is set, 3 seconds
is used for images (as before), or the actual duration for video cards.

Also rename the options to --opening-credits and --closing-credits.  To
add multiple cards, the options must be passed multiple times.  The
nargs='+' mode was conflicting with the positional arguments.